+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Allowed a type of `bool` (not a literal) to be passed to ensure_result typing wise.
 - Some edge cases in the weekly schedule increment handling logic.
+- Long running `delete_after` and command execution tasks will no-longer be cancelled by GC.
 
 ### Removed
 - `default_permission` options for slash commands as command permissions V2 broke and deprecated this.